Move form-related actions to form component
